Remove UK Data protection act
“Personal data” in the UK Data protection act has the same
meaning as it has in the GDPR.

Quote from
https://www.legislation.gov.uk/ukpga/2018/12/section/5/enacted

"Terms used in Chapter 2 of this Part and in the GDPR have
the same meaning in Chapter 2 as they have in the GDPR."

"Terms used in Chapter 3 of this Part and in the applied GDPR have
the same meaning in Chapter 3 as they have in the applied GDPR."

Note that GDPR is a EU regulation. EU regulations are binding
in its entirety and directly applicable in all Member States:
https://en.wikipedia.org/wiki/Regulation_(European_Union)

Therefore it is a waste of time to ask all module maintainers
to check both GDPR and UK Data protection act whether data
processed by a module falls under the definition of "Personal data".

Please remove the reference to UK Data protection act.